game state wip

+++ Assignment1/Packet.cpp
@@ -32,5 +32,5 @@ Packet::Packet
 	: moveType{ type }, position{ pos }
 {
 	strcpy_s(hostName, name);
-	hostNameLength = strlen(hostName);
+	hostNameLength = static_cast<int>(strlen(hostName));
 }
